refactor(desktop): agent 编排域归入 services/agent + 命名对齐
把 agent 编排相关文件归到 services/agent/ 子目录并统一命名（目录已表 agent，去冗余前缀，同 pr-agent 那轮）：
- main/agent-review.ts → services/agent/review.ts：runAgentReview → runReview、AgentReviewDeps → ReviewDeps
- main/agent-planning.ts → services/agent/planning.ts：runAgentPlanning → runPlanning、AgentPlanningDeps → PlanningDeps
- main/agent-labels.ts → services/agent/labels.ts：buildAgentStepLabels → buildStepLabels
- services/agent-orchestrator.ts → services/agent/orchestrator.ts：AgentOrchestratorService → Orchestrator
- 新增 services/agent/index.ts 桶导出：对外只暴露 Orchestrator，review/planning/labels 为域内协作件不外暴露。
- ipc / context 改从 ./services/agent 引用 Orchestrator；orchestrator 内对 review/planning/labels 改为同目录引用。

纯改名 / 移动、行为不变；desktop typecheck/lint/build 全绿。
