+++ apps/desktop/src/main/ipc.ts
@@ -3,7 +3,7 @@ import * as agent from './controllers/agent.js';
 import * as app from './controllers/app.js';
 import * as config from './controllers/config.js';
 import * as pr from './controllers/pr.js';
-import { AgentOrchestratorService } from './services/agent-orchestrator.js';
+import { Orchestrator } from './services/agent/index.js';
 import {
   createServiceContext,
   setControllerContext,
@@ -30,7 +30,7 @@ export function registerIpcHandlers(deps: RegisterDeps): {
   // run 队列：pragent:run（PR 域）、Agent 编排、AutoPilot 三方共用。
   const runQueue = new RunQueue(base);
   // Agent 编排：复用 run 队列派发工具 run（agent 低优先级泳道）。
-  const orchestrator = new AgentOrchestratorService(base, runQueue);
+  const orchestrator = new Orchestrator(base, runQueue);
   // controller 层统一上下文：基础上下文 + 两个跨域 service，安装为进程级单例（controller 经 getContext() 取用）。
   const ctx: ControllerContext = { ...base, runQueue, orchestrator };
   setControllerContext(ctx);
